requirements.txt
python-dotenv
openai
